Updated Go, AWS SDK, added go modules, moved code to pkg
