+++ .gitignore
@@ -2,3 +2,4 @@ vendor
 bin
 *.sock
 .DS_Store
+*.swp
