+++ go.mod
@@ -1,0 +1,17 @@
+module github.com/kubernetes-sigs/aws-encryption-provider
+
+go 1.12
+
+require (
+	github.com/aws/aws-sdk-go v1.19.37
+	github.com/gogo/protobuf v1.0.0 // indirect
+	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/golang/protobuf v1.0.0 // indirect
+	github.com/stretchr/testify v1.3.0 // indirect
+	golang.org/x/net v0.0.0-20180314180242-e0c57d8f86c1 // indirect
+	golang.org/x/sync v0.0.0-20190423024810-112230192c58 // indirect
+	golang.org/x/text v0.3.0 // indirect
+	google.golang.org/genproto v0.0.0-20180306020942-df60624c1e9b // indirect
+	google.golang.org/grpc v1.10.0
+	k8s.io/apiserver v0.0.0-20180315093817-627fa76a8b04
+)
