+++ pkg/cloud/mock.go
