+++ pkg/plugin/plugin_test.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kubernetes-sigs/aws-encryption-provider/cloud"
+	"github.com/kubernetes-sigs/aws-encryption-provider/pkg/cloud"
 	pb "k8s.io/apiserver/pkg/storage/value/encrypt/envelope/v1beta1"
 )
