+++ test/integration/integration_test.go
@@ -18,10 +18,10 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kubernetes-sigs/aws-encryption-provider/cloud"
-	"github.com/kubernetes-sigs/aws-encryption-provider/connection"
-	"github.com/kubernetes-sigs/aws-encryption-provider/plugin"
-	"github.com/kubernetes-sigs/aws-encryption-provider/server"
+	"github.com/kubernetes-sigs/aws-encryption-provider/pkg/cloud"
+	"github.com/kubernetes-sigs/aws-encryption-provider/pkg/connection"
+	"github.com/kubernetes-sigs/aws-encryption-provider/pkg/plugin"
+	"github.com/kubernetes-sigs/aws-encryption-provider/pkg/server"
 	pb "k8s.io/apiserver/pkg/storage/value/encrypt/envelope/v1beta1"
 )
